# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: R&D Intelligence Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/dashboard:
    get:
      operationId: getDashboardStats
      tags: [analytics]
      summary: Get dashboard KPI stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /analytics/trends:
    get:
      operationId: getTrends
      tags: [analytics]
      summary: Get project and formulation trends
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trends data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrendsData"
  /analytics/cost-simulation:
    post:
      operationId: runCostSimulation
      tags: [analytics]
      summary: Simulate cost impact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostSimulationRequest"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostSimulationResult"
  /analytics/ai-suggestions:
    post:
      operationId: getAiSuggestions
      tags: [analytics]
      summary: Get AI formulation suggestions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSuggestionsRequest"
      responses:
        "200":
          description: AI suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSuggestionsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, scientist, analyst, viewer]
        department:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        stage:
          type: string
          enum: [ideation, research, formulation, testing, validation, scale_up, commercialization]
        status:
          type: string
          enum: [active, on_hold, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, critical]
        leadId:
          type: integer
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        targetDate:
          type: string
          format: date-time
          nullable: true
        successRate:
          type: number
          nullable: true
        revenueImpact:
          type: number
          nullable: true
        productCategory:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        lead:
          $ref: "#/components/schemas/User"
          nullable: true
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - stage
        - status
        - priority
        - tags
        - taskCount
        - completedTaskCount
        - createdAt
...
    DashboardStats:
      type: object
      properties:
        totalProjects:
          type: integer
        activeProjects:
          type: integer
        completedProjects:
          type: integer
        successRate:
          type: number
        avgTimeToMarket:
          type: number
        totalRevenueImpact:
          type: number
        totalFormulations:
          type: integer
        approvedFormulations:
          type: integer
        teamSize:
          type: integer
        projectsByStage:
          type: array
          items:
            $ref: "#/components/schemas/StageCount"
        projectsByStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        monthlyProjects:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyData"
        recentProjects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
      required:
        - totalProjects
        - activeProjects
        - completedProjects
        - successRate
        - avgTimeToMarket
        - totalRevenueImpact
        - totalFormulations
        - approvedFormulations
        - teamSize
        - projectsByStage
        - projectsByStatus
        - monthlyProjects
        - recentProjects
    StageCount:
      type: object
      properties:
        stage:
          type: string
        count:
          type: integer
      required:
        - stage
        - count
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count
    MonthlyData:
      type: object
      properties:
        month:
          type: string
        projects:
          type: integer
        formulations:
          type: integer
      required:
        - month
        - projects
        - formulations
    TrendsData:
      type: object
      properties:
        successByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategorySuccess"
        sensoryCorrelation:
          type: array
          items:
            $ref: "#/components/schemas/SensoryCorrelation"
        costTrends:
          type: array
          items:
            $ref: "#/components/schemas/CostTrend"
        formulationSuccessRate:
          type: number
      required:
        - successByCategory
        - sensoryCorrelation
        - costTrends
        - formulationSuccessRate
    CategorySuccess:
      type: object
      properties:
        category:
          type: string
        successRate:
          type: number
        count:
          type: integer
      required:
        - category
        - successRate
        - count
    SensoryCorrelation:
      type: object
      properties:
        formulationId:
          type: integer
        name:
          type: string
        overallScore:
          type: number
        successRate:
          type: number
      required:
        - formulationId
        - name
        - overallScore
        - successRate
    CostTrend:
      type: object
      properties:
        month:
          type: string
        avgCost:
          type: number
        avgMargin:
          type: number
      required:
        - month
        - avgCost
        - avgMargin
    CostSimulationRequest:
      type: object
      properties:
        formulationId:
          type: integer
        ingredientChanges:
          type: array
          items:
            $ref: "#/components/schemas/IngredientChange"
      required:
        - formulationId
        - ingredientChanges
    IngredientChange:
      type: object
      properties:
        ingredientName:
          type: string
        costChangePercent:
          type: number
      required:
        - ingredientName
        - costChangePercent
    CostSimulationResult:
      type: object
      properties:
        originalCost:
          type: number
        newCost:
          type: number
        costChange:
          type: number
        costChangePercent:
          type: number
        originalMargin:
          type: number
        newMargin:
          type: number
        marginChange:
          type: number
        recommendations:
          type: array
          items:
            type: string
      required:
        - originalCost
        - newCost
        - costChange
        - costChangePercent
        - originalMargin
        - newMargin
        - marginChange
        - recommendations
    AiSuggestionsRequest:
      type: object
      properties:
        projectId:
          type: integer
        targetProfile:
          type: string
        constraints:
          type: string
      required:
        - projectId
    AiSuggestionsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/FormulationSuggestion"
        predictedSuccessRate:
          type: number
        reasoning:
          type: string
      required:
        - suggestions
        - predictedSuccessRate
        - reasoning
    FormulationSuggestion:
      type: object
      properties:
        ingredientName:
          type: string
        suggestedPercentage:
          type: number
        rationale:
          type: string
      required:
        - ingredientName
        - suggestedPercentage
        - rationale
...