# lib/db/src/schema/users.ts
import { pgTable, serial, text, boolean, timestamp, pgEnum } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const userRoleEnum = pgEnum("user_role", [
  "admin",
  "manager",
  "ceo",
  "npd_technologist",
  "head_of_product_development",
  "head_of_department",
  "key_account_manager",
  "senior_key_account_manager",
  "project_manager",
  "procurement",
  "scientist",
  "analyst",
  "hr",
  "quality_control",
  "graphics_designer",
  "viewer"
]);

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  email: text("email").notNull().unique(),
  name: text("name").notNull(),
  passwordHash: text("password_hash").notNull(),
  role: userRoleEnum("role").notNull().default("viewer"),
  department: text("department"),
  jobPosition: text("job_position"),
  phone: text("phone"),
  country: text("country"),
  avatar: text("avatar"),
  isActive: boolean("is_active").notNull().default(true),
  createdAt: timestamp("created_at").notNull().defaultNow(),
  updatedAt: timestamp("updated_at").notNull().defaultNow(),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({ id: true, createdAt: true, updatedAt: true });
export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;